Get Endpoints setup
